🌐 lang: 完成自定义 CDN 页、自定义文件名页翻译

+++ src/gui/dialog/setting/custom_cdn_host.py
@@ -1,4 +1,5 @@
 import wx
+import gettext
 
 from utils.config import Config
 from utils.common.thread import ThreadPoolExecutor
@@ -7,9 +8,11 @@ from utils.module.web.ping import Ping
 
 from gui.component.window.dialog import Dialog
 
+_ = gettext.gettext
+
 class CustomCDNDialog(Dialog):
     def __init__(self, parent):
-        Dialog.__init__(self, parent, "自定义 CDN 节点")
+        Dialog.__init__(self, parent, _("自定义 CDN 节点"))
 
         self.init_UI()
 
@@ -20,18 +23,18 @@ class CustomCDNDialog(Dialog):
         self.CenterOnParent()
 
     def init_UI(self):
-        cdn_lab = wx.StaticText(self, -1, "CDN 节点列表")
+        cdn_lab = wx.StaticText(self, -1, _("CDN 节点列表"))
 
         self.cdn_list = wx.ListCtrl(self, -1, size = self.FromDIP((670, 250)), style = wx.LC_REPORT | wx.LC_SINGLE_SEL)
 
-        custom_lab = wx.StaticText(self, -1, "自定义")
+        custom_lab = wx.StaticText(self, -1, _("自定义"))
         self.custom_box = wx.TextCtrl(self, -1, size = self.FromDIP((240, -1)))
-        self.add_btn = wx.Button(self, -1, "添加", size = self.get_scaled_size((80, 28)))
-        self.delete_btn = wx.Button(self, -1, "删除", size = self.get_scaled_size((80, 28)))
+        self.add_btn = wx.Button(self, -1, _("添加"), size = self.get_scaled_size((80, 28)))
+        self.delete_btn = wx.Button(self, -1, _("删除"), size = self.get_scaled_size((80, 28)))
         self.up_btn = wx.Button(self, -1, "↑", size = self.get_scaled_size((28, 28)))
         self.down_btn = wx.Button(self, -1, "↓", size = self.get_scaled_size((28, 28)))
 
-        self.ping_btn = wx.Button(self, -1, "Ping 测试", size = self.get_scaled_size((100, 28)))
+        self.ping_btn = wx.Button(self, -1, _("Ping 测试"), size = self.get_scaled_size((100, 28)))
 
         action_hbox = wx.BoxSizer(wx.HORIZONTAL)
         action_hbox.Add(custom_lab, 0, wx.ALL & (~wx.TOP) & (~wx.BOTTOM) | wx.ALIGN_CENTER, self.FromDIP(6))
@@ -45,8 +48,8 @@ class CustomCDNDialog(Dialog):
 
         bottom_line = wx.StaticLine(self, -1)
 
-        self.ok_btn = wx.Button(self, wx.ID_OK, "确定", size = self.get_scaled_size((80, 30)))
-        self.cancel_btn = wx.Button(self, wx.ID_CANCEL, "取消", size = self.get_scaled_size((80, 30)))
+        self.ok_btn = wx.Button(self, wx.ID_OK, _("确定"), size = self.get_scaled_size((80, 30)))
+        self.cancel_btn = wx.Button(self, wx.ID_CANCEL, _("取消"), size = self.get_scaled_size((80, 30)))
 
         bottom_hbox = wx.BoxSizer(wx.HORIZONTAL)
         bottom_hbox.AddStretchSpacer(1)
@@ -72,22 +75,22 @@ class CustomCDNDialog(Dialog):
 
     def init_utils(self):
         def init_listctrl():
-            self.cdn_list.AppendColumn("CDN 节点", width = self.FromDIP(350))
-            self.cdn_list.AppendColumn("优先级", width = self.FromDIP(120))
-            self.cdn_list.AppendColumn("延迟", width = self.FromDIP(100))
+            self.cdn_list.AppendColumn(_("优先级"), width = self.FromDIP(120))
+            self.cdn_list.AppendColumn(_("CDN 节点"), width = self.FromDIP(350))
+            self.cdn_list.AppendColumn(_("延迟"), width = self.FromDIP(100))
 
         def init_cdn_list():
             for cdn in Config.Temp.cdn_list:
-                self.cdn_list.Append([cdn, "", "未检测"])
+                self.cdn_list.Append(["", cdn, _("未检测")])
 
-            self.setItemOrder()
+            self.update_priority()
 
         init_listctrl()
         init_cdn_list()
 
     def onOKEVT(self):
         if self.cdn_list.GetItemCount() == 0:
-            wx.MessageDialog(self, "保存失败\n\n至少需要添加一个 CDN 节点", "警告", wx.ICON_WARNING).ShowModal()
+            wx.MessageDialog(self, _("保存失败\n\n至少需要添加一个 CDN 节点"), _("警告"), wx.ICON_WARNING).ShowModal()
             return
         
         Config.Temp.cdn_list = [self.cdn_list.GetItemText(i, 0) for i in range(self.cdn_list.GetItemCount())]
@@ -96,8 +99,8 @@ class CustomCDNDialog(Dialog):
         def worker(index: int, cdn: str):
             def update(value):
                 self.cdn_list.SetItem(index, 2, value)
-            
-            wx.CallAfter(update, "正在检测...")
+
+            wx.CallAfter(update, _("正在检测..."))
 
             result = Ping.run(cdn)
 
@@ -112,69 +115,69 @@ class CustomCDNDialog(Dialog):
 
     def onAddEVT(self, event):
         if not self.custom_box.GetValue():
-            wx.MessageDialog(self, "添加失败\n\n请输入要添加的 CDN 节点", "警告", wx.ICON_WARNING).ShowModal()
+            wx.MessageDialog(self, _("添加失败\n\n请输入要添加的 CDN 节点"), _("警告"), wx.ICON_WARNING).ShowModal()
             self.custom_box.SetFocus()
             return
 
         for i in range(self.cdn_list.GetItemCount()):
-            if self.cdn_list.GetItemText(i, 0) == self.custom_box.GetValue():
-                wx.MessageDialog(self, "添加失败\n\n无法重复添加已存在的项目", "警告", wx.ICON_WARNING).ShowModal()
+            if self.cdn_list.GetItemText(i, 1) == self.custom_box.GetValue():
+                wx.MessageDialog(self, _("添加失败\n\n无法重复添加已存在的项目"), _("警告"), wx.ICON_WARNING).ShowModal()
                 self.custom_box.SetFocus()
                 return
-                    
-        index = self.cdn_list.Append([self.custom_box.GetValue(), "", "未检测"])
 
-        self.setItemOrder()
+        index = self.cdn_list.Append([str(self.cdn_list.GetItemCount() + 1), self.custom_box.GetValue(), _("未检测")])
 
         self.cdn_list.Select(index)
         self.cdn_list.Focus(index)
 
     def onDeleteEVT(self, event):
-        dlg = wx.MessageDialog(self, "删除项目\n\n确定要删除选定的项目吗？", "警告", wx.ICON_WARNING | wx.YES_NO)
+        if self.check_item_focused():
+            return
+
+        dlg = wx.MessageDialog(self, _("删除项目\n\n确定要删除选定的项目吗？"), _("警告"), wx.ICON_WARNING | wx.YES_NO)
 
         if dlg.ShowModal() == wx.ID_YES:
             self.cdn_list.DeleteItem(self.cdn_list.GetFocusedItem())
 
     def onUpEVT(self, event):
-        current_item = self.cdn_list.GetFocusedItem()
-
-        if current_item > 0:
-            previous_item = current_item - 1
-
-            host = self.cdn_list.GetItemText(previous_item, 0)
-            latency = self.cdn_list.GetItemText(previous_item, 2)
-
-            self.cdn_list.DeleteItem(previous_item)
-            self.cdn_list.InsertItem(current_item, "")
-
-            self.cdn_list.SetItem(current_item, 0, host)
-            self.cdn_list.SetItem(current_item, 2, latency)
-
-            self.setItemOrder()
-
-            self.cdn_list.Select(previous_item)
-            self.cdn_list.Focus(previous_item)
-
-    def onDownEVT(self, event):
-        current_item = self.cdn_list.GetFocusedItem()
+        if self.check_item_focused():
+            return
         
-        if current_item < self.cdn_list.GetItemCount() - 1:
-            next_item = current_item + 1
+        self.move_to(self.cdn_list.GetFocusedItem() - 1)
 
-            host = self.cdn_list.GetItemText(next_item, 0)
-            latency = self.cdn_list.GetItemText(next_item, 2)
+    def onDownEVT(self, event):   
+        if self.check_item_focused():
+            return
+             
+        self.move_to(self.cdn_list.GetFocusedItem() + 1)
 
-            self.cdn_list.DeleteItem(next_item)
-            self.cdn_list.InsertItem(current_item, "")
+    def move_to(self, index: int):
+        current_item = self.cdn_list.GetFocusedItem()
 
-            self.cdn_list.SetItem(current_item, 0, host)
-            self.cdn_list.SetItem(current_item, 2, latency)
+        if index in range(0, self.cdn_list.GetItemCount()):
+            label = self.cdn_list.GetItemText(current_item, 1)
+            latency = self.cdn_list.GetItemText(current_item, 2)
 
-            self.setItemOrder()
+            self.cdn_list.DeleteItem(current_item)
+            self.cdn_list.InsertItem(index, "")
 
-            self.cdn_list.Select(next_item)
-            self.cdn_list.Focus(next_item)
+            self.cdn_list.SetItem(index, 1, label)
+            self.cdn_list.SetItem(index, 2, latency)
 
-    def setItemOrder(self):
+            self.cdn_list.Select(index)
+            self.cdn_list.Focus(index)
+        else:
+            wx.Bell()
+
+        self.cdn_list.SetFocus()
+
+        self.update_priority()
+
+    def update_priority(self):
         for index in range(self.cdn_list.GetItemCount()):
-            self.cdn_list.SetItem(index, 1, str(index + 1))
+            self.cdn_list.SetItem(index, 0, str(index + 1))
+
+    def check_item_focused(self):
+        if self.cdn_list.GetFocusedItem() == -1:
+            wx.MessageDialog(self, _("未选择项目\n\n请选择要调整的项目"), _("警告"), wx.ICON_WARNING).ShowModal()
+            return True
